# api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: korotkov.ivan.s@gmail.com
    name: Ivan Korotkov
    url: https://github.com/iskorotkov
  description: A simple manager for storing your passwords
  title: Password manager
  version: 1.0.0
servers:
- url: /
paths:
  /api/v1/passwords:
    get:
      description: Get list of passwords
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Password'
                type: array
                x-content-type: application/json
          description: Success
        "500":
          description: Internal server error
    post:
      description: Create a new password
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Password'
      responses:
        "201":
          description: Success
        "400":
          description: Invalid password submitted
        "500":
          description: Internal server error
  /api/v1/passwords/{id}:
    delete:
      description: Delete a password
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          description: Password deleted
        "500":
          description: Internal server error
    get:
      description: Get a password by its ID
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Password'
          description: Success
        "404":
          description: No password with this ID
        "500":
          description: Internal server error
    put:
      description: Update a password
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Password'
      responses:
        "201":
          description: Success
        "400":
          description: Invalid password submitted
        "404":
          description: No password with this ID
        "500":
          description: Internal server error
components:
  schemas:
    Password:
      description: Username and password for a website
      example:
        website: https://github.com
        password: MySecurePassword
        id: 1
        username: admin
      properties:
        id:
          type: integer
        website:
          type: string
        username:
          type: string
        password:
          type: string
      type: object