# go/model_password.go
/*
 * Password manager
 *
 * A simple manager for storing your passwords
 *
 * API version: 1.0.0
 * Contact: korotkov.ivan.s@gmail.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

// Password - Username and password for a website
type Password struct {
	Id int32 `json:"id,omitempty"`

	Website string `json:"website,omitempty"`

	Username string `json:"username,omitempty"`

	Password string `json:"password,omitempty"`
}

// AssertPasswordRequired checks if the required fields are not zero-ed
func AssertPasswordRequired(obj Password) error {
	return nil
}

// AssertRecursePasswordRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of Password (e.g. [][]Password), otherwise ErrTypeAssertionError is thrown.
func AssertRecursePasswordRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aPassword, ok := obj.(Password)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertPasswordRequired(aPassword)
	})
}